+++ src/headless/plugins/adhoc/api.js
@@ -71,7 +71,10 @@ export default {
                     <command sessionid="${sessionid}" node="${node}" action="${action}" xmlns="${Strophe.NS.ADHOC}">
                         ${ !['cancel', 'prev'].includes(action) ? stx`
                             <x xmlns="${Strophe.NS.XFORM}" type="submit">
-                                ${ inputs.map(({ name, value }) => stx`<field var="${name}"><value>${value}</value></field>`) }
+                                ${ inputs.map(({ name, value }) => stx`<field var="${name}">
+                                    ${ Array.isArray(value)
+                                        ? value.map(v => stx`<value>${v}</value>`) : stx`<value>${value}</value>` }
+                                </field>`) }
                             </x>` : '' }
                     </command>
                 </iq>`;
@@ -91,9 +94,10 @@ export default {
 
             const command = result.querySelector('command');
             const status = command?.getAttribute('status');
+            const resultData = command?.querySelector('x[type=result]');
             return {
                 status,
-                ...(status === 'executing' ? parseCommandResult(result) : {}),
+                ...(status === 'executing' || (status === 'completed' && resultData) ? parseCommandResult(result) : {}),
                 note: result.querySelector('note')?.textContent
             }
         }