+++ src/headless/types/shared/types.d.ts
@@ -70,6 +70,7 @@ export type XFormField = {
     type?: XFormFieldTypes;
     text?: string;
     value?: string;
+    values?: string[];
     required?: boolean;
     checked?: boolean;
     options?: XFormOption[];