+++ src/types/plugins/chatview/chat.d.ts
@@ -52,7 +52,7 @@ declare const ChatView_base: {
         enableUpdating(_requestedUpdate: boolean): void;
         attributeChangedCallback(name: string, _old: string | null, value: string | null): void;
         __propertyToAttribute: any;
-        requestUpdate(name?: PropertyKey, oldValue?: unknown, options?: import("lit").PropertyDeclaration): void;
+        requestUpdate(name?: PropertyKey, oldValue?: unknown, options?: import("lit").PropertyDeclaration, useNewValue?: boolean, newValue?: unknown): void;
         __enqueueUpdate: any;
         scheduleUpdate(): void | Promise<unknown>;
         performUpdate(): void;