+++ src/types/plugins/controlbox/controlbox.d.ts
@@ -53,7 +53,7 @@ declare const ControlBoxView_base: {
         enableUpdating(_requestedUpdate: boolean): void;
         attributeChangedCallback(name: string, _old: string | null, value: string | null): void;
         __propertyToAttribute: any;
-        requestUpdate(name?: PropertyKey, oldValue?: unknown, options?: import("lit").PropertyDeclaration): void;
+        requestUpdate(name?: PropertyKey, oldValue?: unknown, options?: import("lit").PropertyDeclaration, useNewValue?: boolean, newValue?: unknown): void;
         __enqueueUpdate: any;
         scheduleUpdate(): void | Promise<unknown>;
         performUpdate(): void;