+++ src/types/shared/chat/toolbar.d.ts
@@ -18,6 +18,9 @@ export class ChatToolbar extends CustomElement {
         show_emoji_button: {
             type: BooleanConstructor;
         };
+        show_location_button: {
+            type: BooleanConstructor;
+        };
         show_send_button: {
             type: BooleanConstructor;
         };
@@ -32,6 +35,7 @@ export class ChatToolbar extends CustomElement {
     show_spoiler_button: boolean;
     show_call_button: boolean;
     show_emoji_button: boolean;
+    show_location_button: boolean;
     connectedCallback(): void;
     render(): import("lit-html").TemplateResult<1>;
     firstUpdated(): void;