+++ CHANGES.md
@@ -12,7 +12,9 @@
 - #3863: Shift `unescapeHTML` helper function (used to set `isOnlyEmoji` on a message) to converse-headless
 - #3889: MUC join: Use room jids localpart as name in case name or identity not found
 - #3916: Add support for XEP-0461 Message Replies, allowing users to reply to specific messages
+- #3939: Don't show invitations to groupchats in which the user is already present
 - Don't render unfurls for retracted messages.
+- #3941: add adhoc completed command result and text-multi as merged lines of text
 - #3949: Allow pinning bookmarked conversations to the top (XEP-0469)
 
 ## 12.0.0 (2025-08-28)